Fixed condition stringifying issue
The FCM server rejects conditions which are surround in single quotes.
When the condition is stringified over the wire, double quotes in it
get converted to \" which the FCM server does not properly handle. We
can get around this by replacing internal double quotes with single
quotes.

This totally seems like an FCM server bug (which I will raise with
them), but for the time being, the workaround is easy enough to
implement.

Change-Id: I9fe5ef98fcc6a917243dd78a915816b1c629f9de